voice-service/agent/requirements.txt
livekit-agents>=0.8.0,<0.12
livekit-plugins-deepgram>=0.6.0
livekit-plugins-openai>=0.8.0
livekit-plugins-silero>=0.6.0
python-dotenv>=1.0.0
httpx>=0.27.0
openai>=1.30.0
zoneinfo; python_version < "3.9"
